refactor: プリセットを API router で切り出して整理
